🏗️ Architect: Modularize optics package and consolidate utilities
- Moved pure mathematical optics functions from `apts/utils/optics.py` to `apts/optics/calculations.py`.
- Updated all internal imports in `apts/optics/` and its submodels to use the new module.
- Maintained backward compatibility for `apts.utils.optics` via redirection in `apts/utils/__init__.py`.
- Removed the redundant `apts/utils/optics.py` file.
- Verified all optics-related unit tests pass.
