+++ apts/optics/models/photometry.py
@@ -2,7 +2,7 @@ import math
 from typing import Optional, TYPE_CHECKING
 import numpy
 from ...units import get_unit_registry
-from ...utils import optics as optics_utils
+from .. import calculations as optics_utils
 
 if TYPE_CHECKING:
     from pint import Quantity
